Updated post install script

+++ post_install.sh
@@ -3,21 +3,47 @@
 # Remove uneeded programs
 echo "Uninstalling unwanted programs..."
 
-sudo pacman -R \
+pacman -R \
 	epiphany
 
 # Install applications
 echo "Instaling applications..."
-sudo pacman -S \
-	i3 \
+sudo pacman -Sy --needed \
 	dmenu \
 	xorg-server \
 	xorg-xinit \
-	xorg-util \
-	yay \
+	xorg-xrandr \
+	i3 \
 	stow \
 	neovim \
-	feh
+	feh \
+	kitty \
+	firefox
 
-echo "Copying xinitrc"
-cp ./xinitrc ~/.xinitrc
+
+# Install yay
+# Check if yay has already been installed 
+if ! command -v yay &> /dev/null
+then
+	echo "Installing yay"
+	pacman -S --needed git base-devel
+	git clone https://aur.archlinux.org/yay.git
+	cd yay
+	makepkg -si
+	cd ..
+else
+	echo "Skipping, yay already installed"
+fi
+
+
+# TODO stow all of the config files
+
+
+# Copy xinitrc if it does not exist
+if ! test -f ~/.xinitrc
+then
+	echo "Copying xinitrc"
+	cp ./xinitrc ~/.xinitrc
+else
+	echo "xinitrc already copied"
+fi
